spikes/thoughtful_ai_test/pyproject.toml
[project]
name = "thoughtful-ai-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.20.1",
    "langchain>=0.3.20",
    "langchain-openai>=0.3.8",
    "openai>=1.66.2",
    "pip>=25.0.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
